Rename WebSocketProvider to WebSocketManager, implement refactor
